gateway-api/pyproject.toml
[project]
name = "gateway-api"
version = "0.1.0"
description = "Clinical Data Gateway API - Provides GP Connect APIs over the internet via the API Management platform"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">3.13,<4.0.0"

[tool.poetry.dependencies]
clinical-data-common = { git = "https://github.com/NHSDigital/clinical-data-common.git", tag = "v0.1.0" }
flask = "^3.1.2"
types-flask = "^1.1.6"
requests = "^2.32.5"

[tool.poetry]
packages = [{include = "gateway_api", from = "src"},
            {include = "stubs", from = "stubs"},
            {include = "fhir", from = "src"}]

[tool.coverage.run]
relative_files = true
omit = ["*/tests/*", "*/features/*", "*/test_*.py"]

[tool.coverage.paths]
source = [
    "src/",
    "*/gateway-api/src/",
]
root = [
    ".",
    "*/gateway-api/",
]

[tool.coverage.xml]
output = "coverage.xml"

[tool.coverage.html]
directory = "htmlcov"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "mypy (>=1.18.2,<2.0.0)",
    "pytest>=8.0.0",
    "pytest-bdd (>=8.1.0,<9.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-html (>=4.1.1,<5.0.0)",
    "pact-python>=2.0.0",
    "python-dotenv>=1.0.0",
    "schemathesis>=4.4.1",
    "types-requests (>=2.32.4.20250913,<3.0.0.0)",
    "types-pyyaml (>=6.0.12.20250915,<7.0.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
]

[tool.mypy]
strict = true
